refactor: 모델 파일들 구조 변경
